% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{check_pandoc_version}
\alias{check_pandoc_version}
\title{Check pandoc version (deprecated)}
\usage{
check_pandoc_version()
}
\value{
stops R Markdown report from running if <2.0
}
\description{
Check pandoc version (deprecated)
}
\examples{
\dontrun{check_pandoc_version()}

}
