% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{detect_namespaces_in_file}
\alias{detect_namespaces_in_file}
\title{Detect package names referenced with :: in a file}
\usage{
detect_namespaces_in_file(path)
}
\arguments{
\item{path}{Path to a text file (for example an Rmd or R file).}
}
\value{
Character vector of unique package names (may be length 0).
}
\description{
Read a file and return package names referenced with :: notation.
}
\seealso{
Other utilities: 
\code{\link{detect_namespaces_in_text}()}
}
\concept{utilities}
