% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_functions.R
\name{.try_paths}
\alias{.try_paths}
\title{Try different combinations of roots and folders for directories that exist}
\usage{
.try_paths(root, path)
}
\arguments{
\item{root}{root to check. joined to path options.}

\item{path}{path to check. joined to root options.

Used in get_server_path() to check all combos of alt_root and alt_path.}
}
\value{
vector of root path combinations that exist or if none exist, exists with stop message
}
\description{
Try different combinations of roots and folders for directories that exist
}
\examples{
\dontrun{.try_paths("N:/", "folder_location")}
}
