% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_functions.R
\name{get_os}
\alias{get_os}
\title{Get operating system}
\usage{
get_os()
}
\value{
one of "linux" "windows" or "osx"
}
\description{
Get operating system
}
\details{
Replaces .Platform$OS.type which only return "windows" or "unix".
}
\examples{
\dontrun{get_os()}
}
