% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{get_output_type}
\alias{get_output_type}
\title{Get output type for warnings and markup}
\usage{
get_output_type()
}
\value{
either 'latex' or 'pandoc'
}
\description{
Use this to set options for warnings or markup when knitting both
PDF (Latex) and Word document reports.
}
\examples{
\dontrun{get_output_type()}

}
