% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_functions.R
\name{get_root}
\alias{get_root}
\title{Get root path based on operating system}
\usage{
get_root(os)
}
\arguments{
\item{os}{can use get_os() to return one of "linux", "osx" or "windows" to return relevant root path}
}
\value{
string for root path to try based on operating system
}
\description{
Get root path based on operating system
}
\examples{
get_root(get_os()) or get_root("windows") or get_root("fake root")
}
