% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_functions.R
\name{get_server_path}
\alias{get_server_path}
\title{Declare Directories to Load in for Unix/Mac versus Windows Operating System}
\usage{
get_server_path(
  folder = c("networks", "trials"),
  alt_root = "",
  alt_folder = "",
  force_windows_path
)
}
\arguments{
\item{folder}{string for either networks or trials. defaults to networks}

\item{alt_root}{string alternative root to find networks or trials file location.}

\item{alt_folder}{string alternative folder mapping to find networks or trials file location.}

\item{force_windows_path}{string forces a windows path}
}
\value{
defaults to "networks" path based on OS.
}
\description{
Declare Directories to Load in for Unix/Mac versus Windows Operating System
}
\examples{
install_git(file.path(get_server_path(),'cavd', 'Studies', 'cvdNNN', 'pdata', 'VDCNNN.git'))
}
