% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{insert_break}
\alias{insert_break}
\title{Insert a page break}
\usage{
insert_break()
}
\value{
Inserts a page break
}
\description{
Use this function in your Rmd document to create a page break across PDF or
Word output types
}
