% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_template.R
\name{insert_ref}
\alias{insert_ref}
\title{Cross-reference a figure or table}
\usage{
insert_ref(ref)
}
\arguments{
\item{ref}{character string with the reference type and chunk label}
}
\value{
inserts a link to a table or figure in a PDF or Word report
}
\description{
Cross-reference a figure or table
}
\examples{

\dontrun{insert_ref("tab:my-results")}
\dontrun{insert_ref("fig:box-plots")}
}
