% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_git_workaround.R
\name{install_git_workaround}
\alias{install_git_workaround}
\title{Wrapper to use \code{\link[remotes:install_git]{remotes::install_git()}} on network drive paths}
\usage{
install_git_workaround(repo_path, ...)
}
\arguments{
\item{repo_path}{Path to \code{.git} repo of (data)package to install}

\item{...}{Additional arguments passed to \code{\link[remotes:install_git]{remotes::install_git()}}}
}
\value{
Called for package installation side effect
}
\description{
Wrapper to use \code{\link[remotes:install_git]{remotes::install_git()}} on network drive paths
}
