% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{install_load_cran_packages}
\alias{install_load_cran_packages}
\title{Check for package, install if needed, and load}
\usage{
install_load_cran_packages(packages)
}
\arguments{
\item{packages}{character vector of package names}
}
\value{
packages will be loaded into environment
}
\description{
Check for package, install if needed, and load
}
\examples{
\dontrun{load_install_cran_packages(c("tidyr", "dplyr"))}
}
