% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{networks_path}
\alias{networks_path}
\title{Construct networks path}
\usage{
networks_path(...)
}
\arguments{
\item{...}{additional path components passed to \code{\link[=file.path]{file.path()}}; appended after the networks root path}
}
\value{
a character vector of the concatenated path
}
\description{
Requires environment variable \code{VISCTEMPLATES_NETWORKS_PATH} to be defined in
your \code{.Renviron} file. This will typically be \verb{/networks} on Linux, \verb{N:} on
Windows, and \verb{/Volumes/networks} on macOS, but may vary depending on your
system's custom mapping. You can edit this file in Rstudio with
\code{\link[usethis:edit]{usethis::edit_r_environ()}}.
}
