% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_obj_compare_fun.R
\name{obj_compare}
\alias{obj_compare}
\title{Compare two data objects}
\usage{
obj_compare(data_object_1, data_object_2, simplify = FALSE)
}
\arguments{
\item{data_object_1}{data object to compare}

\item{data_object_2}{data object to compare against}

\item{simplify}{Should the "differences" output contain detailed comparisons? Defaults to FALSE}
}
\value{
a list indicating the differences between the two data frames. If no differences are found, a character string is returned.
}
\description{
Compare two data objects
These are objects that are in environment,
NOT with git commits.
}
\section{Last updated by}{

Valeria Duran
}

\section{Last updated date}{

01/11/2023
}

\examples{

df<-data.frame(x=c(1:10),y=c(6:15))
df2<-data.frame(x=c(1:10), y=c(15:24))
obj_compare(df,df2)
obj_compare(df,df2, simplify = TRUE)
obj_compare(df,df)
}
