% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_obj_compare_fun.R
\name{object_compare}
\alias{object_compare}
\title{Compare data objects from two different commits.}
\usage{
object_compare(dataObject = NULL, SHA1 = NULL, SHA2 = NULL)
}
\arguments{
\item{SHA1}{first commit id}

\item{SHA2}{second commit id}

\item{dataobject}{the data object name}
}
\value{
a list of:
\item{

}
}
\description{
Compare data objects from two different commits.
}
\details{
working directory must be pointed to repo clone
}
\examples{
Lusso847
repoList <- object_compare(dataObject = "nab", SHA1 = "f199ca8" , SHA2 = "30aa5d5")
}
