% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{print_visc_table}
\alias{print_visc_table}
\title{Print a pretty table for both PDF and Word}
\usage{
print_visc_table(
  df,
  caption,
  caption_short = NULL,
  label = NULL,
  fontsize = 8,
  digits = 3,
  longtable = FALSE,
  longtable_clean_cut = TRUE,
  col_widths_inches = NULL,
  collapse_rows = NULL,
  bold_rows = NULL,
  header = NULL,
  footer_list = NULL,
  vlines_after = NULL,
  hlines_after = NULL
)
}
\arguments{
\item{df}{data frame to be printed as a table}

\item{caption}{caption (long version)}

\item{caption_short}{optional short caption for table of conetents. if NULL (default), caption argument is used.}

\item{label}{label to use for table (alternative to setting label in chunk options)}

\item{fontsize}{default is 8}

\item{digits}{default is 3}

\item{longtable}{TRUE or FALSE}

\item{longtable_clean_cut}{Used by kableExtra::collapse_rows, default is TRUE}

\item{col_widths_inches}{Numeric list of column widths in inches.
Specify columns in order starting with the first and continuing up to the last column that requires a specified width.
An NA at position i in the list means that the width of column i will not be changed from the default.}

\item{collapse_rows}{list of column numbers to collapse across rows}

\item{bold_rows}{list of row numbers to make bold}

\item{header}{Use kableExtra::add_header_above style}

\item{footer_list}{list of footnotes  (list of strings)}

\item{vlines_after}{list of column numbers to after which to insert vertical lines}

\item{hlines_after}{list of row numbers to after which to insert horizontal lines}
}
\description{
Prints a pretty and consistently formatted table for both PDF and Word
outputs using the same function. Uses kable for PDF outputs, and flextable
for Word and pandoc.
}
