% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{set_pandoc_markup}
\alias{set_pandoc_markup}
\title{Set pandoc markup}
\usage{
set_pandoc_markup(output_type)
}
\arguments{
\item{output_type}{character string of document output type}
}
\value{
logical
}
\description{
Use this for conditionally formatting output when knitting both
PDF (Latex) and Word document reports.
}
\examples{
\dontrun{

pandoc_markup <- set_markup_warnings(output_type = get_output_type())

my_results |>
  kable() |>
  cell_spec(pvalue, bold = ifelse(pandoc_markup, TRUE, FALSE))

}
}
