% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{trials_path}
\alias{trials_path}
\title{Construct trials path}
\usage{
trials_path(...)
}
\arguments{
\item{...}{additional path components passed to \code{\link[=file.path]{file.path()}}; appended after
the trials root path}
}
\value{
a character vector of the concatenated path
}
\description{
Requires environment variable \code{VISCTEMPLATES_TRIALS_PATH} to be defined in
your \code{.Renviron} file. To do this,
\enumerate{
\item Edit this file in Rstudio with \code{\link[usethis:edit]{usethis::edit_r_environ()}}
\item Add a line to \code{.Renviron} defining \code{VISCTEMPLATES_TRIALS_PATH}.
Only use the forward slash \code{/} as a path separator and do not use a trailing \code{/}.
Typical settings by operating system:
\itemize{
\item Windows
\preformatted{
VISCTEMPLATES_TRIALS_PATH="T:"
}
\item macOS
\preformatted{
VISCTEMPLATES_TRIALS_PATH="/Volumes/trials"
}
\item Linux
\preformatted{
VISCTEMPLATES_TRIALS_PATH="/trials"
}
}
\item Save the file
\item Restart your R session
}
}
