% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_slide_deck.R
\name{use_slide_deck}
\alias{use_slide_deck}
\title{Use a Slide Deck}
\usage{
use_slide_deck(
  deck_name = "VDCnnn_assay_slides",
  path = ".",
  interactive = TRUE
)
}
\arguments{
\item{deck_name}{name of the file/folder (character, no extension, no subdirectory)}

\item{path}{folder within the active project where the deck should be created}

\item{interactive}{TRUE by default; FALSE for non-interactive unit testing only}
}
\description{
Creates a template R Markdown file for a slide presentation (PPTX)
}
\examples{
\dontrun{
use_slide_deck(
  deck_name = "VDCnnn_Assay_Slides",
  path = "slides"
)
}
}
