% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_visc_project.R
\name{use_visc_docs}
\alias{use_visc_docs}
\title{Create a VISC docs directory with template files}
\usage{
use_visc_docs(study_name)
}
\arguments{
\item{study_name}{name of study in VDCNNN format}
}
\description{
Creates the docs/ directory and presentations/ directory with templates
for the project background, objectives, group colors, study schema, and
project-level bib file.
}
\examples{
\dontrun{
use_visc_docs("Gallo477")
}
}
