% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_visc_gitignore.R
\name{use_visc_gitignore}
\alias{use_visc_gitignore}
\title{Add to .gitignore file}
\usage{
use_visc_gitignore(directory = ".")
}
\arguments{
\item{directory}{Directory relative to active project to set ignores.
Defaults to current project directory.}
}
\description{
Adds common files that are ignored in VISC statistical reports.
}
\examples{
\dontrun{
use_visc_gitignore()
}
}
