% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_visc_report.R
\name{use_visc_methods}
\alias{use_visc_methods}
\title{Use template files for methods sections in PT reports}
\usage{
use_visc_methods(
  path = ".",
  assay = c("generic", "bama", "nab", "adcc"),
  interactive = TRUE,
  reproducibility
)
}
\arguments{
\item{path}{path within the active project}

\item{assay}{"generic", "bama", "nab" or "adcc"}

\item{interactive}{TRUE by default. FALSE is for non-interactive unit testing
only.}

\item{reproducibility}{FALSE by default. TRUE adds reproducibility methods to
the end of stat-methods}
}
\description{
Creates a "methods" directory that contains 3 "child" R Markdown documents
used in PT reports: statistical-methods.Rmd, lab-methods.Rmd,
and biological-endpoints.Rmd
}
\examples{
\dontrun{
use_visc_methods(path = "bama/BAMA-PT-Report", assay = "bama")
}
}
