% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_visc_pr_template.R
\name{use_visc_pr_template}
\alias{use_visc_pr_template}
\title{Add a pull request template to a hidden .github directory}
\usage{
use_visc_pr_template()
}
\value{
saves .md template for use on GitHub
}
\description{
Add a pull request template to a hidden .github directory
}
