% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{use_visc_protocol}
\alias{use_visc_protocol}
\title{Create a VISC protocol/ directory with template files}
\usage{
use_visc_protocol(study_name)
}
\arguments{
\item{study_name}{name of study in VDCNNN format}
}
\description{
Creates the protocol/ directory and presentations/ directory with templates
for the project background, objectives, group colors, and study schema.
}
\examples{
\dontrun{
use_visc_readme("Gallo477")
}
}
