% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_visc_project.R
\name{use_visc_readme}
\alias{use_visc_readme}
\title{Use a VISC README template}
\usage{
use_visc_readme(study_name, save_as = "README.Rmd")
}
\arguments{
\item{study_name}{name of study in VDCNNN format}

\item{save_as}{where to save README.Rmd. Defaults to top-level.}
}
\description{
Use a VISC README template
}
\examples{
\dontrun{
use_visc_readme("Gallo477")
}
}
