% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_visc_report.R
\name{use_visc_report}
\alias{use_visc_report}
\title{Use a VISC Report Template}
\usage{
use_visc_report(
  report_name = "VDCnnn_assay_PTreport",
  path = ".",
  report_type = c("empty", "generic", "bama", "nab", "adcc"),
  interactive = TRUE,
  reproducibility = FALSE
)
}
\arguments{
\item{report_name}{name of the file (character)}

\item{path}{path of the file within the active project}

\item{report_type}{"empty", "generic", "bama", "nab", or "adcc"}

\item{interactive}{TRUE by default. FALSE is for non-interactive unit testing
only}

\item{reproducibility}{FALSE by default. TRUE adds reproducibility methods to
the end of stat-methods}
}
\description{
This function creates a template R Markdown file for a VISC report.
}
\examples{
\dontrun{
use_visc_report(
  report_name = "McElrath708_BAMA_PT_Report_blinded",
  path = "BAMA",
  report_type = "bama"
  )
}
}
