% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_visc_report.R
\name{use_visc_report_test_suite}
\alias{use_visc_report_test_suite}
\title{Use PT report QC test template for a specfic report}
\usage{
use_visc_report_test_suite(report_name, path = ".", interactive = TRUE)
}
\arguments{
\item{report_name}{name of the file (character)}

\item{path}{path of the file within the active project}

\item{interactive}{TRUE by default. FALSE is for non-interactive unit testing
only.}
}
\description{
Creates two unit test files in tests/testthat/ that check code and outputs
for path/report_name. Tests can be run with devtools::test().
}
\examples{
\dontrun{
use_visc_report_test_suite(report_name = "VDCnnn_BAMA_PTreport", path = "BAMA")
}
}
