% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_visc_project.R
\name{use_visc_startup}
\alias{use_visc_startup}
\title{Use a VISC study startup checklist template for the project}
\usage{
use_visc_startup(study_name)
}
\arguments{
\item{study_name}{name of study in VDCNNN format}
}
\description{
Use a VISC study startup checklist template for the project
}
\examples{
\dontrun{
use_visc_startup("Gallo477")
}
}
