% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{visc_compile}
\alias{visc_compile}
\title{Knit VISC report to network folder}
\usage{
visc_compile(
  input = visc_compile_input(),
  output_dir = visc_compile_output_dir(input)
)
}
\arguments{
\item{input}{Path to Rmd file to knit. Defaults to currently open Rstudio editor tab}

\item{output_dir}{Path to write PDF, log, and ancillary knit files. Defaults to same relative path in VISC project as currently open Rstudio editor tab}
}
\value{
The directory to which files were written
}
\description{
Knit VISC report to network folder
}
