% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{visc_compile_input}
\alias{visc_compile_input}
\title{Helper to detect \code{\link[=visc_compile]{visc_compile()}} default input file.}
\usage{
visc_compile_input()
}
\value{
The normalized path of the currently open Rstudio editor tab, or an error
}
\description{
Helper to detect \code{\link[=visc_compile]{visc_compile()}} default input file.
}
