% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{visc_compile_output_dir}
\alias{visc_compile_output_dir}
\title{Helper to compute \code{\link[=visc_compile]{visc_compile()}} default output directory.}
\usage{
visc_compile_output_dir(input_file)
}
\arguments{
\item{input_file}{Input file from which to compute the output directory}
}
\value{
relative output directory for \code{\link[=visc_compile]{visc_compile()}}. Intended to be \if{html}{\out{<project>}}/.../<this_report_folder>/
}
\description{
Helper to compute \code{\link[=visc_compile]{visc_compile()}} default output directory.
}
