% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{visc_pdf_document}
\alias{visc_pdf_document}
\title{Convert to a VISC Report PDF/LaTeX document}
\usage{
visc_pdf_document(latex_engine = "pdflatex", keep_tex = TRUE, ...)
}
\arguments{
\item{latex_engine}{latex engine to use}

\item{keep_tex}{keep the .tex file}

\item{...}{other options to pass to \code{bookdown::pdf_document2()}}
}
\description{
Runs the VISC Report for PDF output based on the template.tex file.
}
\details{
Normally used through \code{output:VISCtemplates::visc_pdf_document} in the .rmd YAML
}
