% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{visc_report}
\alias{visc_report}
\title{Runs VISC Report (PDF Output)}
\usage{
visc_report(latex_engine = "pdflatex", keep_tex = TRUE)
}
\arguments{
\item{latex_engine}{latex engine to use}

\item{keep_tex}{keep the .tex file}
}
\description{
Runs the VISC Report for PDF output based on the template.tex file.
}
\details{
Normally used through `output:VISCtemplates::visc_report` in the .rmd YAML
}
