% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_visc_project.R
\name{create_visc_project}
\alias{create_visc_project}
\title{Create a VISC project}
\usage{
create_visc_project(path)
}
\arguments{
\item{path}{file path}
}
\value{
opens a new RStudio session with template project directory
}
\description{
Create a VISC project
}
