% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{insert_break}
\alias{insert_break}
\title{Insert a page break (deprecated)}
\usage{
insert_break()
}
\value{
Inserts a page break (deprecated)
}
\description{
Do not use in new code. Use latex command
\code{\\newpage} in your Rmd instead. This function is retained to not
break old Rmd files created from previous templates.
}
