% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{insert_break}
\alias{insert_break}
\title{Insert a page break}
\usage{
insert_break()
}
\value{
inserts a page break
}
\description{
Use this function in an R Markdown document to insert a page break when you
are knitting both PDF (Latex) and Word document reports.
}
\examples{
\dontrun{visc_break()}
}
