% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{networks_path}
\alias{networks_path}
\title{Construct networks path}
\usage{
networks_path(...)
}
\arguments{
\item{...}{additional path components passed to \code{\link[=file.path]{file.path()}}; appended after
the networks root path}
}
\value{
a character vector of the concatenated path
}
\description{
Requires environment variable \code{VISCTEMPLATES_NETWORKS_PATH} to be defined (on
its own line) in your \code{.Renviron} file. To make changes, edit this file in
Rstudio with \code{\link[usethis:edit]{usethis::edit_r_environ()}}, save the file, then restart your R
session. Only use the forward slash \code{/} as a path separator and do not use a
trailing \code{/}. Typical settings by operating system:
\itemize{
\item Windows
\preformatted{
VISCTEMPLATES_NETWORKS_PATH="N:"
}
\item macOS
\preformatted{
VISCTEMPLATES_NETWORKS_PATH="/Volumes/networks"
}
\item Linux
\preformatted{
VISCTEMPLATES_NETWORKS_PATH="/networks"
}
}
}
