% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{trials_path}
\alias{trials_path}
\title{Construct trials path}
\usage{
trials_path(...)
}
\arguments{
\item{...}{character vectors; path components passed to \code{\link[=file.path]{file.path()}}}
}
\value{
a character vector of the concatenated path
}
\description{
Requires environment variable VISCTEMPLATES_TRIALS_PATH to be defined in
your .Renviron file. You can edit this file in Rstudio with
\code{\link[usethis:edit]{usethis::edit_r_environ()}}.
}
