% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{trials_path}
\alias{trials_path}
\title{Construct trials path}
\usage{
trials_path(...)
}
\arguments{
\item{...}{additional path components passed to \code{\link[=file.path]{file.path()}}; appended after the trials root path}
}
\value{
a character vector of the concatenated path
}
\description{
Requires environment variable \code{VISCTEMPLATES_TRIALS_PATH} to be defined in
your \code{.Renviron} file. This will typically be \verb{/trials} on Linux, \verb{T:} on
Windows, and \verb{/Volumes/trials} on macOS, but may vary depending on your
system's custom mapping. You can edit this file in Rstudio with
\code{\link[usethis:edit]{usethis::edit_r_environ()}}.
}
