% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{trials_path}
\alias{trials_path}
\title{Construct trials path}
\usage{
trials_path(...)
}
\arguments{
\item{...}{additional path components passed to \code{\link[=file.path]{file.path()}}; appended after
the trials root path}
}
\value{
a character vector of the concatenated path
}
\description{
Requires environment variable \code{VISCTEMPLATES_TRIALS_PATH} to be defined (on
its own line) in your \code{.Renviron} file. To make changes, edit this file in
Rstudio with \code{\link[usethis:edit]{usethis::edit_r_environ()}}, save the file, then restart your R
session. Do not include trailing path separator like \code{/} or \verb{\\}. Typical
settings by operating system:
\itemize{
\item Windows
\preformatted{
VISCTEMPLATES_TRIALS_PATH="T:"
}
\item macOS
\preformatted{
VISCTEMPLATES_TRIALS_PATH="/Volumes/trials"
}
\item Linux
\preformatted{
VISCTEMPLATES_TRIALS_PATH="/trials"
}
}
}
