% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{use_visc_methods}
\alias{use_visc_methods}
\title{Use template files for methods sections in PT reports}
\usage{
use_visc_methods(path = ".", assay = c("generic", "bama", "nab", "adcc"))
}
\arguments{
\item{path}{path within the active project}

\item{assay}{"generic", "bama", "nab" or "adcc"}
}
\description{
Creates a "methods" directory that contains 3 "child" R Markdown documents
used in PT reports: statistical-methods.Rmd, lab-methods.Rmd,
and biological-endpoints.Rmd
}
\examples{
\dontrun{
use_visc_methods(path = "bama/BAMA-PT-Report", assay = "bama")
}
}
