% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{use_visc_methods}
\alias{use_visc_methods}
\title{Use template files for methods sections in PT reports}
\usage{
use_visc_methods(assay = c("generic", "bama"), directory = ".")
}
\arguments{
\item{assay}{"bama" or "generic"}

\item{directory}{folder to save files}
}
\value{

}
\description{
Creates three "child" R Markdown documents used in PT reports:
statistical-methods.Rmd, lab-methods.Rmd, and biological-endpoints.Rmd
}
\examples{
#' \dontrun{
use_visc_methods(assay = "bama", directory = "BAMA/BAMA_PT_Report")
}
}
