% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{use_visc_report}
\alias{use_visc_report}
\title{Use a VISC Report Template}
\usage{
use_visc_report(
  report_name = "PT-Report",
  path = ".",
  report_type = c("empty", "generic", "bama")
)
}
\arguments{
\item{report_name}{name of the file (character)}

\item{path}{path of the file within the active project}

\item{report_type}{"empty", "generic", or "bama"}
}
\description{
Use a VISC Report Template
}
\examples{
\dontrun{
use_visc_report(report_name = "BAMA-PT-Report", path = "bama", report_type = "bama")
}
}
