% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visc_load_pdata.R
\name{visc_load_pdata}
\alias{visc_load_pdata}
\title{Load a VISC pdata object and check version by datapackage or hash}
\usage{
visc_load_pdata(.data, proj_or_datapackage = "datapackage", criteria = NULL)
}
\arguments{
\item{.data}{pdata as name or character, e.g., PKGNAME_ASSAY or "PKGNAME_ASSAY"}

\item{proj_or_datapackage}{whether to load the data from the current project
repo or an installed datapackage}

\item{criteria}{32 digit hash or data package version 0.1.X.}
}
\value{
pdata object
}
\description{
Allows for loading a pdata object either from active project
repo (during review) or from the library installed location. This
facilitates task switching when transitioning from ad hoc review to
production reporting.
}
\examples{
\dontrun{
# default behavior: loads a pdata from the currently installed package, no criteria checking
visc_load_pdata(Hassell750_ics)

## add a check against hash
visc_load_pdata(Hassell750_ics, criteria = "4f054442a6549bffcd947af4b0da9155")
## add a check against dataversion
visc_load_pdata(Hassell750_ics, criteria = "0.1.68")


# load a pdata from the active project repo, not installed
visc_load_pdata(Hassell750_ics, proj_or_datapackage = "proj")

## add check against hash
visc_load_pdata(Hassell750_ics, proj_or_datapackage = "proj",
  criteria = "09ab8a5a3831e854d21144d89557ccb1")
## skips criteria check if looking at datapackage
}
}
