% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visc_load_pdata.R
\name{visc_load_pdata}
\alias{visc_load_pdata}
\title{Load a VISC pdata object and check data hash}
\usage{
visc_load_pdata(
  .data,
  proj_or_datapackage = c("datapackage", "proj", "repo"),
  criteria = NULL,
  package = NULL,
  lib.loc = NULL
)
}
\arguments{
\item{.data}{pdata as name or character, e.g., PKGNAME_ASSAY or
"PKGNAME_ASSAY"}

\item{proj_or_datapackage}{whether to load the data from an installed
datapackage (\code{datapackage}, the default) or the current project repo
(\code{proj} or \code{repo})}

\item{criteria}{character, a 32-digit hexadecimal data hash with lowercase
letters}

\item{package}{only used in \code{datapackage} mode. NULL (default; standard VISC
pdata naming) or character. If NULL, look for pdata named \code{.data} in
package <portion of \code{.data} before the first underscore>. If not NULL, look
for pdata named \code{.data} in package \code{package}.}

\item{lib.loc}{library path from which to load the data package. Passed
internally to \code{utils::data()}. Default is \code{NULL}, which uses the first
element of \code{.libPaths()}}
}
\value{
pdata object
}
\description{
Allows for loading a pdata object either from active project
repo (e.g., during review) or from the library installed location. This
facilitates task switching when transitioning from ad hoc review to
production reporting.
}
\examples{
\dontrun{
# default behavior: loads a pdata from the currently installed package, no criteria checking
visc_load_pdata(Hassell750_ics)

## add a check against hash
visc_load_pdata(Hassell750_ics, criteria = "4f054442a6549bffcd947af4b0da9155")

# load a pdata from the active project repo, not installed
visc_load_pdata(Hassell750_ics, proj_or_datapackage = "proj")

## add check against hash
visc_load_pdata(Hassell750_ics, proj_or_datapackage = "proj",
  criteria = "09ab8a5a3831e854d21144d89557ccb1")
## skips criteria check if looking at datapackage
}
}
