% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_slide_deck.R
\name{visc_pptx_document}
\alias{visc_pptx_document}
\title{Custom PowerPoint output format for VISC slide decks}
\usage{
visc_pptx_document(style = c("Light", "Navy"), ...)
}
\arguments{
\item{style}{Character, one of c("Light", "Navy"). Selects
which Fred Hutch branded PPTX template to use as \code{reference_doc}.}

\item{...}{Additional arguments passed to
\code{\link[rmarkdown:powerpoint_presentation]{rmarkdown::powerpoint_presentation()}}, e.g. \code{slide_level}.}
}
\description{
Wraps \code{\link[rmarkdown:powerpoint_presentation]{rmarkdown::powerpoint_presentation()}}, defaulting \code{reference_doc}
to one of the Fred Hutch branded PPTX templates bundled with VISCtemplates.
}
